requirements.txt
discord.py[voice]
pynacl
python-dotenv
ping3
pytz